Split routes and notifcation channels and use proper filename
